Update the page-front.tpl.php and page.tpl.php
Change the structure of the page templates files of the theme to include the
footer section on all the files to render the page properly. The <script> code
was also deleted from all the pages to be loaded from the carpross.js and have
a much clenaer html structure without any javascript code.
This still pending the cleaning of the indentation.

	modified:   templates/page--front.tpl.php
	modified:   templates/page.tpl.php